Delete Script
Added a delete playlists script, its in a folder called DO NOT RUN and, upon running, it prompts the user if they are very sure they want to delete everything, then it forces the user to login to the spotify account and copy the resulting url into the terminal. So those three safeguards should keep it pretty safe.